Define a protocol that includes view/over/set functionality
Because lenses, prisms, and traversals will all implement the
view/over/set functionality, this moves these functions to a protocol.

Compose is currently in the protocol, but to facilitate composing
lenses, prisms, and traversals with one another, it will likely be moved
out and implemented in a top level module.