Add function heads for view/set/over to allow pipeline-able versions
Adds additional heads to the main lens functions that allow the order of
the lens and data structure arguments to be reversed. This allows users
to pass arguments in whichever order they prefer, and it enables
better pipeline compatibility for lens operations.

I am slightly opposed to the ambiguity that this can create and might
prefer distinctly named functions. If this becomes an issue later, this
is subject to change.