Fixes creating and composing idx lenses with bad data
If the data structure indexed into was an empty list, the lenses created with
idx/1 would raise an error; this is because the Keyword.keyword? check returns
true for an empty list, so the wrong type of getter was being created. This adds
an explicit check for an empty list and properly handles it if that is the case.

Additionally, using a lens on a non-keyword list that isn't a valid index now
returns {:error, {:lens, :bad_path}}.